pyproject.toml
[project]
name = "product-fidelity-lab"
version = "0.1.0"
description = "Product Photography AI Evaluation System"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "structlog>=24.4",
    "httpx>=0.28",
    "fal-client>=0.5",
    "google-genai>=1.0",
    "diskcache>=5.6",
    "aiosqlite>=0.21",
    "scikit-image>=0.25",
    "scikit-learn>=1.6",
    "numpy>=2.2",
    "Pillow>=11.1",
    "scipy>=1.14",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.25",
    "ruff>=0.9",
    "pyright>=1.1",
]

[project.scripts]
pfl-demo = "product_fidelity_lab.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/product_fidelity_lab"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]

[tool.ruff.lint.per-file-ignores]
"scripts/*.py" = ["E402", "E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (real API calls)",
    "costly: marks tests that cost money",
]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "strict"
include = ["src"]
